requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pymysql
requests
httpx
pandas
python-multipart
openpyxl
numpy
cryptography
aiosmtplib
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
googletrans